rename image-metadata util
