+++ src/_eleventy/utils/image-metadata-index.js
